requirements.txt
pymongo
python-decouple
pandas
numpy
typing
flask
openpyxl
# -e .

